Ints changed to Floaters PutPixel calls changed to covert floats to ints

+++ Engine/Bone.h
@@ -6,16 +6,16 @@
 class Bone
 {
 public:
-	void Init(int in_x, int in_y); 
+	void Init(float in_x, float in_y); 
 	void Update();
 	void ProcessConsumptionForward(const DogForward& dogf);
 	void ProcessConsumptionBackward(const DogBackward& dogb);
 	void Draw(Graphics& gfx) const;
 	bool IsEaten() const;
 private:
-	int x;
-	int y;
-	static constexpr int Width = 12;
-	static constexpr int Height = 7;
+	float x;
+	float y;
+	static constexpr float Width = 12.0f;
+	static constexpr float Height = 7.0f;
 	bool isEaten = false; 
 };
